api/v3alpha: add license_details field to version message
deps.dev currently serves licenses mapped to SPDX expressions, or
"non-standard" if no unambiguous mapping is known.

To better serve users who want to inspect original licenses, add a new
field containing the license both as specified by the package author in
the package metadata (or, for Go, as determined by the licensecheck
package) and mapped to SPDX.

Addresses #62.

Change-Id: I5dc4cb20b8904d75454da98865a5b7c4a8aa72b8