util/maven: replace BoolString with TruthyBool and FalsyBool (#117)
For a string field holding boolean value, the default value may be true,
and may be false. Currently, we do not specify which type of boolean
string, and the client needs to check Maven documentation to decide the
actual boolean value.

This PR replaces the existing `BoolString` with `TruthyBool` (default
value is true) and `FalsyBool` (default value is false), and implements
`Boolean()` method for both types so that the clients no need to check
the Maven documentation for the default value.

Removing `BoolString` is a breaking change, but considering that `util`
packages are not versioned yet, I am not too worried about this.