Add CalculateMinVersion function for NPM. (#269)
This function is relevant for cases where we are given a constraint
range and want to see the minimum version that fulfills that constraint,
for a heuristic/approximation of which version may be pulled at runtime.
This does not look up actual versions, so there is no guarantee that the
version returned here actually exists (e.g. >1.0.0 will return 1.0.1,
which may not be a real version).